# docs/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EduMark",
    "description": "\n基于大模型的中小学作业批改系统，提供以下核心能力：\n\n- **作业提交与批改**：上传作业图片，自动识别题目并进行 AI 批改\n- **知识库管理**：上传教学资料，构建知识库辅助批改\n- **用户管理**：学生/老师/管理员三种角色，权限分级\n- **模型管理**：配置视觉模型和语言模型\n- **配置管理**：个性化识别模式与增强选项\n",
    "contact": {
      "name": "EduMark Team",
      "email": "admin@edumark.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users/{id}": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "查询用户",
        "description": "根据用户 ID 查询用户详细信息，仅管理员可操作。",
        "operationId": "get_user_api_v1_users__id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "用户 ID",
              "title": "Id"
            },
            "description": "用户 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功，返回用户信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_UserOut_"
                }
              }
            }
          },
          "403": {
            "description": "无管理员权限"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "用户管理"
        ],
        "summary": "更新用户",
        "description": "更新用户信息，仅管理员可操作。只需传入需要修改的字段。",
        "operationId": "update_user_api_v1_users__id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "用户 ID",
              "title": "Id"
            },
            "description": "用户 ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              },
              "examples": {
                "更新昵称": {
                  "summary": "修改用户昵称",
                  "value": {
                    "username": "张三（新昵称）"
                  }
                },
                "更新邮箱": {
                  "summary": "修改用户邮箱",
                  "value": {
                    "email": "newemail@example.com"
                  }
                },
                "禁用用户": {
                  "summary": "禁用用户账号",
                  "value": {
                    "is_active": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功，返回用户信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_UserOut_"
                }
              }
            }
          },
          "403": {
            "description": "无管理员权限"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "用户管理"
        ],
        "summary": "删除用户",
        "description": "删除用户，仅管理员可操作。执行以下限制：\n\n- 不能删除自己\n- 不能删除 root 账号\n- 删除时会级联删除该用户的所有作业、知识文档和配置",
        "operationId": "delete_user_api_v1_users__id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "用户 ID",
              "title": "Id"
            },
            "description": "用户 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "403": {
            "description": "无管理员权限"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "新增用户",
        "description": "创建新用户，仅管理员可操作。\n\n- 账号需全局唯一，重复创建会报错\n- 创建用户时会自动初始化默认配置",
        "operationId": "create_user_api_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              },
              "examples": {
                "创建学生用户": {
                  "summary": "创建一个学生账号",
                  "value": {
                    "account": "zhangsan",
                    "username": "张三",
                    "password": "123456",
                    "email": "zhangsan@example.com",
                    "role": "student"
                  }
                },
                "创建老师用户": {
                  "summary": "创建一个老师账号",
                  "value": {
                    "account": "teacher_wang",
                    "username": "王老师",
                    "password": "123456",
                    "email": "wang@example.com",
                    "role": "teacher"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "创建成功，返回用户信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_UserOut_"
                }
              }
            }
          },
          "403": {
            "description": "无管理员权限"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/users/page": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "分页查询用户",
        "description": "分页查询用户列表，仅管理员可操作。\n\n- 支持排序和过滤条件\n- 默认每页 10 条，最大 100 条",
        "operationId": "page_user_api_v1_users_page_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageDTO"
              },
              "examples": {
                "默认分页": {
                  "summary": "默认分页查询（第1页，每页10条）",
                  "value": {
                    "page_num": 1,
                    "page_size": 10
                  }
                },
                "按角色过滤": {
                  "summary": "筛选学生用户",
                  "value": {
                    "page_num": 1,
                    "page_size": 10,
                    "query": {
                      "role": "student"
                    }
                  }
                },
                "按账号排序": {
                  "summary": "按账号升序排列",
                  "value": {
                    "page_num": 1,
                    "page_size": 10,
                    "sort_fields": [
                      {
                        "field": "account",
                        "direction": "asc"
                      }
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "查询成功，返回分页数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_PageVO_UserOut__"
                }
              }
            }
          },
          "403": {
            "description": "无管理员权限"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageDTO": {
        "properties": {
          "page_num": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Num",
            "description": "页码，从 1 开始",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "description": "每页条数，范围 1~100",
            "default": 10
          },
          "sort_fields": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SortField"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Fields",
            "description": "排序规则列表，可选"
          },
          "query": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query",
            "description": "查询过滤条件，键值对形式，可选"
          }
        },
        "type": "object",
        "title": "PageDTO"
      },
...
      "PageVO_UserOut_": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "总记录数"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/UserOut"
            },
            "type": "array",
            "title": "Rows",
            "description": "当前页数据列表"
          }
        },
        "type": "object",
        "required": [
          "total",
          "rows"
        ],
        "title": "PageVO[UserOut]",
        "example": {
          "rows": [],
          "total": 100
        }
      },
...
      "Result": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
...
      "Result_PageVO_UserOut__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageVO_UserOut_"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[PageVO[UserOut]]",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
...
      "Result_UserOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[UserOut]",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
      "SortDirection": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortDirection"
      },
      "SortField": {
        "properties": {
          "field": {
            "type": "string",
            "minLength": 1,
            "title": "Field",
            "description": "排序字段名"
          },
          "direction": {
            "$ref": "#/components/schemas/SortDirection",
            "description": "排序方向，asc=升序，desc=降序",
            "default": "asc"
          }
        },
        "type": "object",
        "required": [
          "field"
        ],
        "title": "SortField"
      },
...
      "UserCreate": {
        "properties": {
          "account": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "title": "Account",
            "description": "用户账号，3~64 个字符"
          },
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "description": "用户昵称，1~64 个字符"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "Password",
            "description": "用户密码，6~128 个字符"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "电子邮箱，可选"
          },
          "role": {
            "type": "string",
            "pattern": "^(student|teacher)$",
            "title": "Role",
            "description": "用户角色：student=学生，teacher=老师",
            "default": "student"
          }
        },
        "type": "object",
        "required": [
          "account",
          "username",
          "password"
        ],
        "title": "UserCreate"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "用户 ID"
          },
          "account": {
            "type": "string",
            "title": "Account",
            "description": "用户账号"
          },
          "username": {
            "type": "string",
            "title": "Username",
            "description": "用户昵称"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "电子邮箱"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "用户角色：student 或 teacher"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "是否启用"
          },
          "create_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Create Time",
            "description": "创建时间"
          },
          "update_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Update Time",
            "description": "最后更新时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "account",
          "username",
          "role",
          "is_active"
        ],
        "title": "UserOut",
        "example": {
          "account": "zhangsan",
          "create_time": "2026-06-01T08:00:00",
          "email": "zhangsan@example.com",
          "id": 1,
          "is_active": true,
          "role": "student",
          "update_time": "2026-06-04T12:00:00",
          "username": "张三"
        }
      },
      "UserUpdate": {
        "properties": {
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "用户昵称"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "pattern": "^[\\w\\.-]+@[\\w\\.-]+\\.\\w+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "电子邮箱"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "description": "是否启用，true=启用，false=禁用"
          }
        },
        "type": "object",
        "title": "UserUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}